components/fastapi-pip-pap/requirements.txt
aiofiles==24.1.0
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.2.3
black==24.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
crypto==1.4.1
cryptography==42.0.5
datamodel-code-generator==0.25.4
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
Flask==3.0.2
genson==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.6
inflect==5.6.2
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
loguru==0.7.2
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
Naked==0.1.32
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pdoc3==0.10.0
platformdirs==4.2.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyOpenSSL==24.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
shellescape==3.8.1
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
uWSGI==2.0.24
Werkzeug==3.0.1
